Update tree structure and new trait
Elements now possess a parent field, which allows traversal to parent elements from children.

New `Find` trait that allows more complex searches for single or multiple elements. In addition, allowing the ability to fetch the value of an element directly or the element itself.

Any methods that returned an `&[T]`, now return `Vec<&T>` instead.

The order of elements retrieved from the `elements` method in metadata is now consistent.